Add preliminary context menu support
This is essentially the widgets part, with some tricks to get it to
honor the widget's context menu policy.
It enables c++11 for the widgets library for the convenience of using lambdas,
which admitedly we could do without, but seems reasonable considering our timeline
and the fact that we build chromium that way.

Change-Id: I6a632a78d2aa48fb0dfecfe491e92651d12407db