Revive QWebEnginePage::evaluateJavascript
We can't have QWebFrame::evaluateJavascript in its old form for
several reasons, the first of which being that we don't have a
QWebEngineFrame class anymore. This is worked around by adding
an optional QString parameter with the XPath.

Another issue is that the WebKit1 API was synchronous, and this would
not play nicely with the very asynchronous nature of chromium we're
now sitting on top of. In order to highlight that when porting, we
need to change the function signature slightly.

This makes evaluateJavaScript a template member function that will
accept function pointers, functors and lambdas,much like the new
signal/slot syntax. QWebEnginePage::IgnoreJavaScriptResult can be
passed when the result is of no interest.

Change-Id: I64e15a6f5a168936c52a4da2cef6285dfd16e0d5